# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for ZuZu.",
        "title": "ZuZu Backend API",
        "contact": {
            "name": "Dima",
            "url": "http://t.me/belozerovmsk"
        }
    },
    "paths": {
...
        "/api/auth/check_auth": {
            "get": {
                "security": [
                    {
                        "AuthKey": []
                    }
                ],
                "description": "Check user is logged in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "CheckAuth",
                "parameters": [
                    {
                        "type": "string",
                        "default": "zuzu-t=xxx",
                        "description": "Token",
                        "name": "Cookie",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "description": "Logout from Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/auth/signin": {
            "post": {
                "description": "Login to Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "SignIn",
                "parameters": [
                    {
                        "description": "SignInPayload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignInPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "400": {
                        "description": "error messege",
                        "schema": {
                            "$ref": "#/definitions/responser.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "description": "Create Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "SignUp",
                "parameters": [
                    {
                        "description": "SignUpPayload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignUpPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "400": {
                        "description": "error messege",
                        "schema": {
                            "$ref": "#/definitions/responser.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
...
        "/api/order/create": {
            "post": {
                "description": "Create Order using profile ID from cookies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "CreateOrder",
                "responses": {
                    "200": {
                        "description": "New order info",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "401": {
                        "description": "User unauthorized"
                    },
                    "404": {
                        "description": "something not found error message",
                        "schema": {
                            "$ref": "#/definitions/responser.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
        "/api/order/get_all": {
            "get": {
                "description": "Get all Orders using profile ID from cookies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "GetOrders",
                "responses": {
                    "200": {
                        "description": "All orders info",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    },
                    "401": {
                        "description": "User unauthorized"
                    },
                    "404": {
                        "description": "something not found error message",
                        "schema": {
                            "$ref": "#/definitions/responser.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
        "/api/order/get_current": {
            "get": {
                "description": "Get Current Order using profile ID from cookies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "GetCurrentOrder",
                "responses": {
                    "200": {
                        "description": "Current order info",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "401": {
                        "description": "User unauthorized"
                    },
                    "404": {
                        "description": "something not found error message",
                        "schema": {
                            "$ref": "#/definitions/responser.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Order": {
            "type": "object",
            "properties": {
                "addressId": {
                    "type": "string"
                },
                "addressIsCurrent": {
                    "type": "boolean"
                },
                "city": {
                    "type": "string"
                },
                "flat": {
                    "type": "string"
                },
                "house": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderProduct"
                    }
                },
                "statusId": {
                    "type": "integer"
                },
                "street": {
                    "type": "string"
                }
            }
        },
        "models.OrderProduct": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "categoryName": {
                    "type": "string"
                },
                "categoryParent": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "productId": {
                    "type": "string"
                },
                "productName": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
...
        "models.Profile": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "models.SignInPayload": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 6
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 8
                }
            }
        },
        "models.SignUpPayload": {
            "type": "object",
            "required": [
                "login",
                "password",
                "phone"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 6
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 8
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "responser.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}